refactor: retire the operator admin curricula tab (#252)
curricula moved to a standalone sidebar screen (#217) already carries
the tab's list/detail/register/reanalyze/delete flows through the
same CurriculaTab component — the admin tab was only kept alive during
the move. Finish it: drop the tab entry, drop the now-unused
`standalone` prop and its SectionHeader branch, and move the folder
out of admin/ so the code layout matches where the screen actually
lives.
