+++ src/features/operator/admin/AdminScreen.tsx
@@ -7,13 +7,14 @@ import { useCohortScope } from './_/cohortScope'
 import { COHORT_STATUS_LABEL } from './_/labels'
 
 /*
-  OP-06 운영 관리 — **조직을 세팅한다.** 기수·반·명단·매니저·교안·비용.
+  OP-06 운영 관리 — **조직을 세팅한다.** 기수·반·명단·매니저·비용.
 
   ▸ **헤더의 기수 스위처가 범위를 정한다.** 셸이 원래 그 자리를 갖고 있었는데(스코프
     선택기) 값이 `7기` 자리값 하나였다 — 실제 목록을 여기서 넘겨 동작하게 만든다.
     화면 안에 또 하나를 두지 않는다: 같은 축이 두 개면 어느 쪽이 이기는지 안 보인다.
-    반·명단·비용이 이 범위를 따르고, **매니저와 교안만 기관 전체다** — 매니저는 기수를
-    옮겨 다니고 교안은 여러 기수가 같이 쓴다.
+    반·명단·비용이 이 범위를 따르고, **매니저만 기관 전체다** — 매니저는 기수를
+    옮겨 다닌다. (교안도 기관 전체 자산이지만 `#217`로 사이드바 최상위 `/operator/curricula`로
+    옮겨져 이제 이 화면 밖에 있다.)
   ▸ **`반`과 `명단`이 갈려 있다.** 반을 만드는 것과 명단을 넣는 것은 한 흐름이지만 두 표가
     한 뷰포트에 안 들어갔다.
   ▸ **오퍼레이터 계정은 여기 없다.** 초대·정지는 슈퍼어드민(SA-02) 소관이고, 이 화면은
