+++ src/features/operator/curricula/CurriculaScreen.route.tsx
