+++ src/features/operator/curricula/CurriculumDetailScreen.route.tsx
