+++ src/features/operator/curricula/CurriculumDetailSkeleton.tsx
