+++ src/features/operator/curricula/components/DeleteCurriculumDialog.tsx
